requirements.txt
streamlit
pandas
numpy
plotly
altair
scikit-learn
snowflake-connector-python
googletrans==4.0.0-rc1
